requirements.txt
altair==5.2.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.1.0
bibtexparser==1.4.1
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.5
django-admin-searchable-dropdown==1.2
django-appconf==1.0.5
django-autocomplete-light==3.9.7
django-cors-headers==4.3.1
django-data-browser==4.2.5
django-filter==23.5
django-modelclone-next==0.8.2
django-select2==8.1.2
django-serverside-datatable==2.1.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
dubo==0.2.9
feedparser==6.0.10
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
iommi==6.7.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Levenshtein==0.23.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mysqlclient==2.2.0
numpy==1.26.2
packaging==23.2
pandas==1.5.3
Pillow==10.0.1
pydeck==0.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzotero==1.5.18
rapidfuzz==3.4.0
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sqlglot==18.17.0
sqlparse==0.4.4
toolz==0.12.0
urllib3==2.0.7
whitenoise==6.6.0
